backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database — requires Python 3.11 (no 3.13 wheel yet)
asyncpg==0.29.0

# Azure OpenAI + HTTP client (versions must be compatible)
openai==1.30.1
httpx==0.27.0


# PDF extraction — text + tables
pdfplumber==0.11.0

# Office formats
python-docx==1.1.0
openpyxl==3.1.2

# Settings
pydantic-settings==2.2.1
python-dotenv==1.0.1